requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.7.0
bitsandbytes>=0.41.0

# Vision-Language Models
pillow>=10.0.0
timm>=0.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
datasets>=2.14.0
pyarrow>=12.0.0

# Experiment tracking and visualization
wandb>=0.15.0
tensorboard>=2.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# API clients for data generation
openai>=1.0.0
anthropic>=0.5.0
requests>=2.31.0

# Configuration and utilities
pyyaml>=6.0
hydra-core>=1.3.0
omegaconf>=2.3.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# Evaluation metrics
scikit-learn>=1.3.0
scipy>=1.11.0

# Model optimization and deployment
onnx>=1.14.0
onnxruntime>=1.16.0
optimum>=1.14.0

# Code quality
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Logging
loguru>=0.7.0

